--- csv/employers.csv
@@ -1,5 +1,0 @@
-﻿Month;1;2;3;4;5;6;7;8;9;10;11;12;13;14;15;16;17;18;19;20;21;22;23;24;25;26;27;28;29;30;31
-Employer name[1];;;4;;;4;4;;4;4;;;4;;2;2;;;2;2;;;;;;4;4;;;4;4
-Employer name[2];1;3;;;1;1;;1;;1;1;;1;4;;;1;1;;1;1;;;;;1;1;;1;;
-.........
-Employer name[n] work_shift;work_shift;work_shift;work_shift;work_shift;work_shift;....work_shift
