--- json/employers_info.json
@@ -1,17 +1,0 @@
-{
-    "Employer_name_1" : {
-        "telegram" : "Employer telegram",
-        "telegram_id" : "Employer telegram id",
-        "group" : "Employer group (ShopMaster/Poisk/CMS/LK)"
-    },
-    "Employer_name_2" : {
-        "telegram" : "Employer telegram",
-        "telegram_id" : "Employer telegram id",
-        "group" : "Employer group (ShopMaster/Poisk/CMS/LK)"
-    },
-    "Employer_name_n" : {
-        "telegram" : "Employer telegram",
-        "telegram_id" : "Employer telegram id",
-        "group" : "Employer group (ShopMaster/Poisk/CMS/LK)"
-    }
-}
