+++ src/sm_bot/__init__.py
@@ -1,0 +1,3 @@
+import os
+
+ROOT_DIR = os.path.dirname(os.path.abspath(__file__))  # root folder
