+++ src/sm_bot/config/config.py
@@ -1,6 +1,7 @@
 import os
 import json
 from dotenv import load_dotenv
+from sm_bot import ROOT_DIR
 
 # Get env variables
 load_dotenv()
@@ -9,11 +10,11 @@ GROUP_CHAT_ID_SM = int(os.environ.get('GROUP_CHAT_ID_SM'))
 GROUP_CHAT_ID_POISK = int(os.environ.get('GROUP_CHAT_ID_POISK'))
 
 # Const
-ROOT_DIR = os.path.dirname(os.path.abspath(__file__))  # root folder
-JSON_DIR_PATH = ROOT_DIR + '/json/'
-CSV_DIR_PATH = ROOT_DIR + '/csv/'
+JSON_DIR_PATH = ROOT_DIR + '/data/json/'
+CSV_DIR_PATH = ROOT_DIR + '/data/csv/'
 
 CSV_PATH = CSV_DIR_PATH + 'employers.csv'
+CSV_PATH_TEST = CSV_DIR_PATH + 'test.csv'
 CSV_PATH_5_2 = CSV_DIR_PATH + 'employers_5_2.csv'
 NEXT_MONTH_CSV_PATH = CSV_DIR_PATH + 'employers-next.csv'
 
@@ -26,3 +27,5 @@ with open(JSON_DIR_PATH + 'employers_month.json', 'r', encoding='utf-8') as mont
 
 with open(JSON_DIR_PATH + 'employers_info.json', 'r', encoding='utf-8') as info_json:
     employers_info = json.load(info_json)
+
+
