+++ src/sm_bot/data/json/employers_info.json
@@ -1,0 +1,46 @@
+{
+    "Employer_name_1" : {
+        "telegram" : "Employer telegram",
+        "telegram_id" : "Employer telegram id",
+        "group" : "Employer group (ShopMaster/Poisk/CMS/LK)",
+        "subscription": {
+            "enabled": bool,
+            "time_to_notify": "20:00"
+        },
+        "webdav": {
+            "name": "Work",
+            "url": "calendar_url",
+            "password": "password"
+        }
+    },
+    "Employer_name_2" : {
+        "telegram" : "Employer telegram",
+        "telegram_id" : "Employer telegram id",
+        "group" : "Employer group (ShopMaster/Poisk/CMS/LK)",
+        "subscription": {
+            "enabled": bool,
+            "time_to_notify": "20:00"
+        },
+        "webdav": {
+            "name": "Work",
+            "url": "calendar_url",
+            "password": "password"
+        }
+    },
+    .........
+    "Employer_name_n" : {
+        "telegram" : "Employer telegram",
+        "telegram_id" : "Employer telegram id",
+        "group" : "Employer group (ShopMaster/Poisk/CMS/LK)",
+        "subscription": {
+            "enabled": bool,
+            "time_to_notify": "20:00"
+        },
+        "webdav": {
+            "name": "Work",
+            "url": "calendar_url",
+            "password": "password"
+        }
+    }
+    
+}
