+++ src/sm_bot/handlers/bot/callback/chatters/__init__.py
@@ -1,0 +1,1 @@
+from sm_bot.handlers.bot.callback.chatters.chatters import *
