+++ src/sm_bot/handlers/bot/callback/subscription/__init__.py
@@ -1,0 +1,1 @@
+from sm_bot.handlers.bot.callback.subscription.subscription import *
