+++ src/sm_bot/handlers/bot/message/base/__init__.py
@@ -1,0 +1,3 @@
+from sm_bot.handlers.bot.message.base.lunch import *
+from sm_bot.handlers.bot.message.base.botinit import *
+from sm_bot.handlers.bot.message.base.log import *
