+++ src/sm_bot/handlers/bot/message/subscription/__init__.py
@@ -1,0 +1,1 @@
+from sm_bot.handlers.bot.message.subscription.subscription import *
