+++ src/sm_bot/handlers/bot/message/webdav/__init__.py
@@ -1,0 +1,1 @@
+from sm_bot.handlers.bot.message.webdav.menu import *
