+++ src/sm_bot/handlers/chattersmanager/__init__.py
@@ -1,0 +1,4 @@
+from sm_bot.handlers.chattersmanager.chatters import Chatters
+from sm_bot.handlers.chattersmanager.chatters import get_lunch_time
+
+today_chatters = Chatters()
