+++ src/sm_bot/handlers/workersmanager/__init__.py
@@ -1,0 +1,3 @@
+from sm_bot.handlers.workersmanager.day_workers import DayWorkers
+
+today_workers = DayWorkers()
