+++ src/sm_bot/services/bot.py
@@ -1,0 +1,4 @@
+import telebot
+import sm_bot.config.config as config
+
+bot = telebot.TeleBot(config.TELEGRAM_TOKEN)
